requirements.txt
pyspellchecker
numpy
language-tool-python
